refactored optics into their own module
